+++ sendparcel_django/__init__.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 
 default_app_config = "sendparcel_django.apps.SendparcelConfig"